backend/modules/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
dnspython==2.4.2
python-whois==0.9.0
nmap==0.7.1
scapy==2.5.0
concurrent-log-handler==0.9.24
requests==2.31.0
beautifulsoup4==4.12.2
python-nmap==0.7.1
python-owasp-zap-v2.4==0.0.14
concurrent-log-handler==0.9.24
pathlib==1.0.1